money 유효성검사 클래스와 에러로 인한 반복 클래스 분리
